+++ rhubarb_lipsync/blender/ui_utils.py
@@ -1,11 +1,11 @@
 import logging
 import pathlib
 import traceback
-from typing import Any, Callable, Iterator, Type
+from typing import Any, Callable, Iterator, Literal, Type
 
 import bpy
 import bpy.utils.previews
-from bpy.types import Area, Context, Sound, UILayout, Window
+from bpy.types import Area, Context, Sound, UILayout, Window, bpy_prop_collection
 
 try:
     from bpy.types import Strip  # Since v4.4
@@ -13,6 +13,9 @@ except ImportError:  # Fall back to old API
     from bpy.types import SoundSequence as Strip
 
 log = logging.getLogger(__name__)
+
+
+OperatorReturnSet = set[Literal['RUNNING_MODAL', 'CANCELLED', 'FINISHED', 'PASS_THROUGH', 'INTERFACE']]
 
 
 def addon_path() -> pathlib.Path:
@@ -187,13 +190,23 @@ def len_limited(iterator: Iterator, max_count=1000) -> int:
     return count
 
 
+def get_strips_from_sequence_editor(context: Context) -> bpy_prop_collection:
+    seq_editor = context.scene.sequence_editor
+    if hasattr(seq_editor, 'sequences'):  # Blender v4
+        return seq_editor.sequences
+    # Blender v5+
+    return seq_editor.strips  # type: ignore
+
+
 def find_sound_strips_by_sound(context: Context, sound: Sound, limit=0) -> list[Strip]:
     '''Finds sound strips which are using the specified sound.'''
     ret: list[Strip] = []
     if not sound or not context.scene.sequence_editor:
         return []
 
-    for i, sq in enumerate(context.scene.sequence_editor.sequences_all):
+    seq_editor = context.scene.sequence_editor
+
+    for i, sq in enumerate(get_strips_from_sequence_editor(context)):
         if limit > 0 and i > limit:
             break  # Limit reached, break the search (for performance reasons)
         if not hasattr(sq, "sound"):